requirements.txt
djangorestframework==3.15.2
django==5.1.1
djangorestframework-simplejwt>=4.8.0